@@ -4,8 +4,8 @@ version: 0.0.3
 homepage: https://github.com/badjio/skeletons/
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: '>=3.2.5 <4.0.0'
+  flutter: 3.22.2
 
 dependencies:
   flutter:
