@@ -78,7 +78,7 @@ class _SkeletonThemeExamplePageState extends State<SkeletonThemeExamplePage> {
               Expanded(
                 child: Text(
                   "Lorem ipsum dolor sit amet.",
-                  style: Theme.of(context).textTheme.headline6,
+                  style: Theme.of(context).textTheme.headlineSmall,
                   maxLines: 1,
                   overflow: TextOverflow.ellipsis,
                 ),
