@@ -150,7 +150,7 @@ class _ThemeModesExamplePageState extends State<ThemeModesExamplePage> {
                 children: [
                   Text(
                     "Lorem ipsum dolor sit amet.",
-                    style: Theme.of(context).textTheme.headline6,
+                    style: Theme.of(context).textTheme.headlineSmall,
                     maxLines: 1,
                     overflow: TextOverflow.ellipsis,
                   ),
